refactor(shared): move list helper to entity
