+++ packages/entity/src/shared/index.ts
@@ -15,6 +15,8 @@ export * from './comment-spec.model';
 export * from './content.model';
 export * from './content.service';
 
+export * from './entity-list-helper';
+
 export * from './event.model';
 export * from './event.service';
 export * from './event-spec.model';
